test: split role tests to separate files
